converting to float32

+++ custom_trainer.py
@@ -147,8 +147,8 @@ class Trainer():
 
 	def _generate_domain_mask(self, domain: int):
 		if domain == 0:
-			return np.full((self.patch_size, self.patch_size), 1.)
-		return np.full((self.patch_size, self.patch_size), 0.)
+			return np.full((self.patch_size, self.patch_size), 1., dtype = 'float32')
+		return np.full((self.patch_size, self.patch_size), 0., dtype = 'float32')
 
 	def train_domain_adaptation(self, patches_dir: list, epochs: int = 25, batch_size: int = 2, val_fraction: float = 0.1,
 								num_images: int = 60, wait: int = 12, rotate: bool = True, flip: bool = True,
@@ -229,7 +229,7 @@ class Trainer():
 				# set lambda value
 				l = lambda_grl(p)
 				self.lambdas.append(l)
-				l = np.full((self.batch_size, 1), l)
+				l = np.full((self.batch_size, 1), l, dtype = 'float32')
 
 				mask = np.asarray([self._generate_domain_mask(domain) for domain in y_classifier_train])
 				step_output = self._training_step_domain_adaptation([x_train, l], [y_segmentation_train, y_classifier_train], mask)
